pyproject.toml
[tool.poetry]
name = "chat-assistance"
version = "0.1.0"
description = ""
authors = ["Rajat Nigam <rajatnigam89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "*"
beautifulsoup4 = "*"
black = "*"
tiktoken = "*"
openai = "*"
pinecone-client = "*"
unstructured = "*"
nltk = "*"
fastapi = "*"
jinja2 = "*"
uvicorn = "*"
streamlit = "*"
streamlit-chat = "*"
tqdm = "*"
poetry-dotenv-plugin = "^0.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
ai = "chat_assistance.ingestion:root"
chat = "chat_assistance.core:root"
gpt = "chat_assistance.gpt-start:root"